requirements.txt
gradio
langchain
langchain_community
langchain_core
langchain_google_genai
python-dotenv
Requests
